Changed the ConstraintLayout to a LinearLayout.
Set the orientation of the layout to vertical.

Set the left, right and top padding to 16dp.

Removed the ID from the layout.

Removed the unneeded xmlns:app declaration.

Removed the "Hello World" TextView.

Added an EditText

Set the ID of the Edit

Set the EditText size to 22sp

Set the width attribute of the EditText to match_parent

Set the height attribute of the EditText to wrap_content

Added a hint to the EditText.

Added a TextView

Added an ID to the TextView.

Set the text size of the TextView to 22sp.

Set the width of the TextView to wrap_content.

Set the height of the TextView to wrap_content.

Set the top margin of the TextView to 8dp.

Added default text to the TextView.

Added a ScrollView

Set the width of the ScrollView to match_parent.

Set the height of the ScrollView to wrap_content.

Set the top margin of the ScrollView to 16dp.

Added a TextView within the ScrollView.

Set the ID of the TextView.

Set the text size of the TextView to 18sp.

Set the height and width of the TextView to wrap_content.

Set the default text of the TextView.